+++ libs/gltfio/include/gltfio/SimpleViewer.h
@@ -138,12 +138,6 @@ public:
     void enableDithering(bool b) { mEnableDithering = b; }
 
     /**
-     * Enables depth prepass on the view.
-     * Defaults to true.
-     */
-    void enablePrepass(bool b) { mEnablePrepass = b; }
-
-    /**
      * Enables FXAA antialiasing in the post-process pipeline.
      * Defaults to true.
      */
@@ -195,7 +189,6 @@ private:
     bool mEnableSunlight = true;
     bool mEnableShadows = true;
     bool mEnableDithering = true;
-    bool mEnablePrepass = true;
     bool mEnableFxaa = true;
     bool mEnableMsaa = true;
     bool mEnableSsao = true;
@@ -417,14 +410,11 @@ void SimpleViewer::updateUserInterface() {
 
     if (ImGui::CollapsingHeader("View")) {
         ImGui::Checkbox("Dithering", &mEnableDithering);
-        ImGui::Checkbox("Depth prepass", &mEnablePrepass);
         ImGui::Checkbox("FXAA", &mEnableFxaa);
         ImGui::Checkbox("MSAA 4x", &mEnableMsaa);
         ImGui::Checkbox("SSAO", &mEnableSsao);
     }
 
-    mView->setDepthPrepass(
-            mEnablePrepass ? View::DepthPrepass::ENABLED : View::DepthPrepass::DISABLED);
     mView->setDithering(mEnableDithering ? View::Dithering::TEMPORAL : View::Dithering::NONE);
     mView->setAntiAliasing(mEnableFxaa ? View::AntiAliasing::FXAA : View::AntiAliasing::NONE);
     mView->setSampleCount(mEnableMsaa ? 4 : 1);